+++ packages/ckeditor5-find-and-replace/tests/findandreplace.js
@@ -135,87 +135,6 @@ describe( 'FindAndReplace', () => {
 			expect( stopSpy.calledOnce ).to.true;
 			expect( stateClearSpy.calledOnce ).to.true;
 		} );
-
-		it( 'should bind FindAndReplaceUI#highlightOffset to FindAndReplaceState#highlightedResult', () => {
-			findAndReplaceEditing.state.highlightedResult = null;
-
-			expect( findAndReplaceUI.highlightOffset ).to.equal( 0 );
-
-			editor.setData( '<p>foo</p>' );
-
-			const firstParagraph = editor.model.document.getRoot().getChild( 0 );
-			let markerA, markerB;
-
-			model.change( writer => {
-				markerA = writer.addMarker( 'findResult:A', {
-					usingOperation: false,
-					affectsData: false,
-					range: writer.createRange(
-						writer.createPositionAt( firstParagraph, 0 ),
-						writer.createPositionAt( firstParagraph, 1 )
-					)
-				} );
-			} );
-
-			model.change( writer => {
-				markerB = writer.addMarker( 'findResult:B', {
-					usingOperation: false,
-					affectsData: false,
-					range: writer.createRange(
-						writer.createPositionAt( firstParagraph, 2 ),
-						writer.createPositionAt( firstParagraph, 3 )
-					)
-				} );
-			} );
-
-			const resultA = {
-				id: 'A',
-				label: 'label',
-				marker: markerA
-			};
-
-			const resultB = {
-				id: 'B',
-				label: 'label',
-				marker: markerB
-			};
-
-			findAndReplaceEditing.state.results.add( resultB, resultA );
-			findAndReplaceEditing.state.highlightedResult = resultB;
-
-			expect( findAndReplaceUI.highlightOffset ).to.equal( 1 );
-		} );
-
-		it( 'should update FindAndReplaceUI#matchCount when FindAndReplaceState#results change', () => {
-			editor.setData( '<p>foo</p>' );
-
-			expect( findAndReplaceUI.matchCount ).to.equal( 0 );
-
-			const firstParagraph = editor.model.document.getRoot().getChild( 0 );
-			let marker;
-
-			model.change( writer => {
-				marker = writer.addMarker( 'findResult:123456', {
-					usingOperation: false,
-					affectsData: false,
-					range: writer.createRange(
-						writer.createPositionAt( firstParagraph, 0 ),
-						writer.createPositionAt( firstParagraph, 1 )
-					)
-				} );
-			} );
-
-			const highlightedResult = {
-				id: '123456',
-				label: 'label',
-				marker
-			};
-
-			findAndReplaceEditing.state.results.add( highlightedResult );
-			findAndReplaceEditing.state.highlightedResult = highlightedResult;
-
-			expect( findAndReplaceUI.matchCount ).to.equal( 1 );
-		} );
 	} );
 
 	describe( 'integration', () => {
